+++ lib/fetching_state.dart
@@ -1,7 +1,5 @@
 library fetching_state;
 
-import 'package:flutter/material.dart';
-
 /// The predefined states when fetching data
 enum FetchingStatus { init, loading, done, error }
 
@@ -50,15 +48,15 @@ class FetchingState<T, E> {
   factory FetchingState.error({E? error}) =>
       FetchingState(fetchingStatus: FetchingStatus.error, error: error);
 
-  /// Declare what which Widget should be return base on [fetchingStatus]
+  /// Declare methods that return [R] base on [fetchingStatus]
   ///
-  /// Allow to skip some method that you don't need or some methods return a same widget,  use [orElse] instead
-  Widget whenOrElse({
-    Widget Function()? onLoading,
-    Widget Function(T? data)? onDone,
-    Widget Function(E? error)? onError,
-    Widget Function()? onInit,
-    required Widget Function()? orElse,
+  /// Allow to use [orElse], skip some methods that you don't need.
+  R whenOrElse<R>({
+    R Function()? onLoading,
+    R Function(T? data)? onDone,
+    R Function(E? error)? onError,
+    R Function()? onInit,
+    required R Function()? orElse,
   }) {
     switch (fetchingStatus) {
       case FetchingStatus.init:
@@ -74,14 +72,15 @@ class FetchingState<T, E> {
     }
   }
 
-  /// Declare which Widget should be return base on [fetchingStatus]
+  /// Declare methods base on [fetchingStatus]
   ///
   /// all state methods are required
-  Widget when({
-    required Widget Function() onLoading,
-    required Widget Function(T? data) onDone,
-    required Widget Function(E? error) onError,
-    required Widget Function() onInit,
+  /// return type [R] need to be the same on all methods
+  R when<R>({
+    required R Function() onLoading,
+    required R Function(T? data) onDone,
+    required R Function(E? error) onError,
+    required R Function() onInit,
   }) {
     switch (fetchingStatus) {
       case FetchingStatus.init: